fix newline bug
When reaching the end of the screen, before this commit you might have
noticed that creating new lines would create two new lines instead of
one.

This wasn't present in earlier commits because the size of struct cell
was 1 char, while with the attr field it has become two.
So `term.cells+sizeof(*term.cells)*term.cols` would evaluate to
`term.cells+2*term.cols` instead of `term.cells+1*term.cols`.
This was never the correct thing to do in the first place because adding
numbers to pointers increments it by value*sizeof(*ptr).