proper line wrapping
We shouldn't be wrapping when setting the cursor position, so this
commit adds a flag that will only allow wrapping if the cursor's
position was not just explicitly set.
I have fixed a couple of bugs that have come about as a result of the
previous commit while I was here because it was making it hard to test.