UTF-8
Now, I'm not entirely sure if the code here is correct, but it works on
the small amount of input I tried it on therefore it works on my
machine.
It may not be fast, but I haven't checked and it's fast enough for me
not to notice, so good enough.
Wide characters aren't supported yet but I am certain the bitmap fonts
don't support them anyway, and I know for a fact they don't support CJK.

To remind me to implement support for wide characters, testdata/utf8 has
been intentionally committed even though the test fails right now.

As an aside, it's a little wasteful to decode UTF-8 only to encode it
again, but fbink doesn't let you print individual runes.