use xkbcommon for input
This commit removes the huge switch statement that I was previously
using to convert evdev keycodes to chars.
The approach was unmaintainable, not to mention that it made it hard for
people with other keyboard layouts (not me) to use inkterm, if they were
to.

xkbcommon takes me 90% of the way and all that's left for me to do is to
handle special keys that would otherwise write escape sequences to the
pty.